Conflicts:
	src/main/java/uy/edu/ort/arqliv/obligatorio/business/aop/UsageAuditAspect.java

+++ src/main/java/uy/edu/ort/arqliv/obligatorio/business/services/ArrivalServiceImpl.java
@@ -41,10 +41,6 @@ public class ArrivalServiceImpl implements ArrivalService {
 	@Autowired
 	IArrivalDAO arrivalDAO;
 	
-//	= (IContainerDAO) ContextSingleton
-//			.getInstance().getBean(PersistenceConstants.ContainerDao);
-//	
-	
 	
 	@Override
 	@Transactional(propagation = Propagation.REQUIRED)
@@ -79,9 +75,6 @@ public class ArrivalServiceImpl implements ArrivalService {
 
 			try {
 			
-//				IArrivalDAO arrivalDAO = (IContainerDAO) ContextSingleton
-//						.getInstance().getBean(PersistenceConstants.ContainerDao);
-				
 				List<Container> containers = new ArrayList<Container>();
 
 				double sumContainerCapacity = 0.0;
@@ -95,7 +88,7 @@ public class ArrivalServiceImpl implements ArrivalService {
 										+ theOperation + " de arribo");
 					}
 
-					if (containerDAO.isContainerInUse(containerId,
+					if (containerDAO.isContainerInUseForArrival(containerId,
 							arrival.getArrivalDate())) {
 						SimpleDateFormat sdfOut = new SimpleDateFormat(
 								"dd/MM/yyyy");
@@ -112,9 +105,6 @@ public class ArrivalServiceImpl implements ArrivalService {
 					containers.add(container);
 					sumContainerCapacity += container.getCapacity();
 				}
-
-//				IShipDAO shipDAO = (IShipDAO) ContextSingleton.getInstance()
-//						.getBean(PersistenceConstants.ShipDao);
 
 				Ship ship = shipDAO.findById(shipId);
 				if (ship == null) {
@@ -139,9 +129,7 @@ public class ArrivalServiceImpl implements ArrivalService {
 				arrival.setShipTransportedWeightThatDay(sumContainerCapacity);
 				arrival.setShip(ship);
 				arrival.setContainers(containers);
-
-//				IArrivalDAO arrivalDAO = (IArrivalDAO) ContextSingleton.getInstance().getBean(PersistenceConstants.ArrivalDao);
-
+ 
 				return arrivalDAO.store(arrival);
 			} catch (CustomServiceException e) {
 				log.error("error " + errorWhen + " el Arribo", e);
@@ -177,15 +165,6 @@ public class ArrivalServiceImpl implements ArrivalService {
 
 		try {
 
-			
-			
-//			IShipDAO shipDAO = (IShipDAO) ContextSingleton.getInstance()
-//					.getBean(PersistenceConstants.ShipDao);
-//			IArrivalDAO arrivalDAO = (IArrivalDAO) ContextSingleton
-//					.getInstance().getBean(PersistenceConstants.ArrivalDao);
-//			IContainerDAO containerDAO = (IContainerDAO) ContextSingleton
-//					.getInstance().getBean(PersistenceConstants.ContainerDao);
-
 			Ship ship = shipDAO.findById(shipId);
 			if (ship == null) {
 				throw new CustomServiceException("el barco con id= " + shipId
@@ -217,7 +196,7 @@ public class ArrivalServiceImpl implements ArrivalService {
 			}
 			
 			if(CONT_CHANGED){
-				Duple<Double, List<Container>> newContainersDuple = getCpacityAndContainers(containerDAO, containerList);
+				Duple<Double, List<Container>> newContainersDuple = getCapacityAndContainers(containerDAO, containerList);
 				double sumContainerCapacity = newContainersDuple.getFirst();
 				List<Container> newContainers = newContainersDuple.getSecond();
 				
@@ -227,13 +206,13 @@ public class ArrivalServiceImpl implements ArrivalService {
 					//cabio contenedor y fecha: no importa si cambio el barco MUST check capacidad sum(cont) < cap barco
 					//AND disponibilidad contenedor en la fecha
 				
-					checAvailability(arrivalDAO, newArrival, originalArrival, 
+					checkAvailability(arrivalDAO, newArrival, originalArrival, 
 							newContainers, false, algoritmoIN);//no hay que sacar al arribo de la lista
 					
 				}else{
 					//cabio contenedor: no importa si cambio el barco MUST check capacidad sum(cont) < cap barco 
 					//AND disponibilidad contenedor en la fecha pero sin tener en cuenta los que ya estan asignados al arribo
-					checAvailability(arrivalDAO, newArrival, originalArrival, 
+					checkAvailability(arrivalDAO, newArrival, originalArrival, 
 							newContainers, true, algoritmoIN);
 				}
 				newArrival.setContainers(newContainers);
@@ -245,7 +224,7 @@ public class ArrivalServiceImpl implements ArrivalService {
 				}
 
 				if(DATE_CHANGED){				
-					checAvailability(arrivalDAO, newArrival, originalArrival, 
+					checkAvailability(arrivalDAO, newArrival, originalArrival, 
 										originalArrival.getContainers(), false, algoritmoIN);
 				}else{
 					//nada que rompa reglas de negocio cambia
@@ -295,7 +274,7 @@ public class ArrivalServiceImpl implements ArrivalService {
 	 * @return
 	 * @throws CustomServiceException
 	 */
-	private boolean checAvailability(IArrivalDAO arrivalDAO, Arrival newArrival, Arrival origArrival,
+	private boolean checkAvailability(IArrivalDAO arrivalDAO, Arrival newArrival, Arrival origArrival,
 			List<Container> containersToCheck, boolean dontCheckAgainstSameArrivalItem, boolean usar1) throws CustomServiceException{
 		
 		
@@ -346,7 +325,7 @@ public class ArrivalServiceImpl implements ArrivalService {
 	 * @return
 	 * @throws CustomServiceException
 	 */
-	private Duple<Double, List<Container>> getCpacityAndContainers(IContainerDAO containerDAO, List<Long> containerList) throws CustomServiceException{
+	private Duple<Double, List<Container>> getCapacityAndContainers(IContainerDAO containerDAO, List<Long> containerList) throws CustomServiceException{
 		List<Container> containers =new ArrayList<Container>();
 		Duple<Double, List<Container>> ret=  new Duple<>(0.0, containers);
 
@@ -439,10 +418,13 @@ public class ArrivalServiceImpl implements ArrivalService {
 	public void delete(String user, long id) throws CustomServiceException {
 		boolean ok = false;
 		try {
-//			IArrivalDAO arrivalDAO = (IArrivalDAO) ContextSingleton
-//					.getInstance().getBean(PersistenceConstants.ArrivalDao);
-
-			ok = arrivalDAO.delete(id);
+			Arrival tmp = arrivalDAO.findById(id);
+			if(!shipDAO.shipInUseInDeparture(tmp.getShip().getId())){
+				ok = arrivalDAO.delete(id);
+			}else{
+				throw new CustomInUseServiceException("No se puede borrar pues está en uso en una partida");	
+			}
+			
 		} catch (Exception e) {
 			log.error("error al dar de baja un Arribo", e);
 			throw new CustomServiceException(e.getMessage(), e);
@@ -455,8 +437,6 @@ public class ArrivalServiceImpl implements ArrivalService {
 	@Override
 	public List<Arrival> list(String user) throws CustomServiceException {
 		try {
-//			IArrivalDAO arrivalDAO = (IArrivalDAO) ContextSingleton
-//					.getInstance().getBean(PersistenceConstants.ArrivalDao);
 			List<Arrival> ret = new ArrayList<Arrival>(arrivalDAO.findAll());
 
 			for (Arrival arrival : ret) {
@@ -483,8 +463,6 @@ public class ArrivalServiceImpl implements ArrivalService {
 	@Override
 	public Arrival find(String user, long id) throws CustomServiceException {
 		try {
-//			IArrivalDAO arrivalDAO = (IArrivalDAO) ContextSingleton
-//					.getInstance().getBean(PersistenceConstants.ArrivalDao);
 
 			Arrival ret = arrivalDAO.findById(id);
 			if (ret != null) {
